fix: set default X01 out rule to single in StandaloneScorer
Changes the initial state of the x01OutRule in the offline StandaloneScorer component from 'double' to 'single' to save casual users a click.

+++ src/components/StandaloneScorer.tsx
@@ -13,7 +13,7 @@ export function StandaloneScorer() {
   const [isPlaying, setIsPlaying] = useState(false);
   const [gameMode, setGameMode] = useState<GameMode>('X01');
   const [startScore, setStartScore] = useState<X01StartScore>(301);
-  const [x01OutRule, setX01OutRule] = useState<'single' | 'double'>('double');
+  const [x01OutRule, setX01OutRule] = useState<'single' | 'double'>('single');
   const [player1Name, setPlayer1Name] = useState('Player 1');
   const [player2Name, setPlayer2Name] = useState('Player 2');
